Fix warnings from external projects, internal warnings and switch back to non-DLL runtime (#723)
Use MultiThreaded and MultiThreadedDebug runtime - not sure how or when it moved away from this
Disable warnings on external projects
Remove Lua compiler and interpreter as we don't use them
Remove unused variables - this fixed warnings in the tests
Add lots of casts
Closes #722

+++ trview.app/Elements/Level.cpp
@@ -93,7 +93,7 @@ namespace trview
 
     uint32_t Level::number_of_rooms() const
     {
-        return _rooms.size();
+        return static_cast<uint32_t>(_rooms.size());
     }
 
     std::vector<Room*> Level::rooms() const
@@ -332,7 +332,7 @@ namespace trview
             {
                 if (sector->flags & SectorFlag::Trigger)
                 {
-                    _triggers.emplace_back(std::make_unique<Trigger>(_triggers.size(), i, sector->x(), sector->z(), sector->trigger()));
+                    _triggers.emplace_back(std::make_unique<Trigger>(static_cast<uint32_t>(_triggers.size()), i, sector->x(), sector->z(), sector->trigger()));
                     room->add_trigger(_triggers.back().get());
                 }
             }
